Updated coverage to include all apps
added test files for all apps

added conftest.py to create fixtures to use across tests in different apss
