+++ Plataforma_de_Vendas/pytest.ini
@@ -1,4 +1,4 @@
 [pytest]
 DJANGO_SETTINGS_MODULE = Plataforma_de_Vendas.settings
 python_files = test_*.py
-addopts = --cov=Accounts --cov-report=term-missing
+addopts = --cov=Accounts --cov=Products --cov=Orders --cov=Stores --cov-report=term-missing
