+++ Plataforma_de_Vendas/tests/conftest.py
@@ -1,6 +1,5 @@
 import pytest
 from rest_framework.test import APIClient
-from django.urls import reverse
 from Accounts.models import CustomUser
 from django.contrib.auth.models import Group
 
@@ -35,29 +34,7 @@ def customer_client(db):
     client.force_authenticate(user=customer_user)
     return client
 
-@pytest.mark.django_db
-class TestGetUsersEndpoint:
-    """Tests access permissions for the all-users-endpoint."""
-
-    url = reverse("all-users-endpoint")
-
-    def test_admin_access(self, admin_client):
-        """Admin should get a 200 OK response."""
-        
-        response = admin_client.get(self.url)
-        assert response.status_code == 200
-
-    def test_seller_access(self, seller_client):
-        """Sellers should get a 401 Unauthorized response."""
-        response = seller_client.get(self.url)
-        assert response.status_code == 401
-
-    def test_unauthenticated_access(self):
-        """Unauthenticated users should get a 403 Forbidden response."""
-        client = APIClient()
-        response = client.get(self.url)
-        assert response.status_code == 403
-    
-
-
-
+@pytest.fixture
+def anonymous_client(db):
+    client = APIClient()
+    return client
